This PR improves the scan speed of Pandas DataFrames by using a zero copy scan for numerics, and avoiding copying ASCII strings by directly accessing the bytes for pure ASCII strings (i.e. `PyUnicode_KIND(val) = PyUnicode_1BYTE_KIND`).


### Performance
Running [this benchmark here](https://gist.github.com/Mytherin/37de933372eb608fb81b9bb64f2df44f) (Q1 on a Pandas DataFrame) this increases performance by a factor 2~ and allows us to run Q1 faster than Pandas does directly on a Pandas DataFrame.

### Old

```
Base Pandas: 2.48 seconds
Pandas Scan: 4.56 seconds
```

### New

```
Base Pandas: 2.17 seconds
Pandas Scan: 2.04 seconds
```